Moved prefs into a use module
